back/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pytz==2024.1
python-dateutil==2.9.0.post0